Moved stores in their own files
